Revert "Add storybook"
