+++ packages/cms/index.ts
@@ -3,4 +3,3 @@ export * from './src/directus'
 export type { HomePage } from './types/home_page'
 export type { StationScreenTranslations } from './types/station_screen_translations'
 export type { TrainLongName } from './types/train_long_name'
-export type { TrainPage } from './types/train_page'
