+++ packages/cms/src/directus.ts
@@ -1,8 +1,7 @@
-import type { Locale } from '../types/common'
 import type { HomePage } from '../types/home_page'
 import type { StationScreenTranslations } from '../types/station_screen_translations'
 import type { TrainLongName } from '../types/train_long_name'
-import type { TrainPage } from '../types/train_page'
+import type { Locale } from '../types/common'
 
 import { camelCaseKeys } from './utils/camel_case_keys'
 
@@ -48,16 +47,8 @@ export const getTrainLongNames = async (
   return json.data.map((longName: TrainLongName) => camelCaseKeys(longName))
 }
 
-export const getTrainPage = async (locale: Locale): Promise<TrainPage> => {
-  const response = await fetch(`${BASE_URL}/items/train_page/${locale}`)
-  const json = await response.json()
-
-  return camelCaseKeys(json.data)
-}
-
 export const directus = {
   getHomePage,
   getStationScreenTranslations,
-  getTrainLongNames,
-  getTrainPage
+  getTrainLongNames
 }
