--- packages/cms/types/train_page.ts
@@ -1,6 +1,0 @@
-import { Locale } from './common'
-
-export interface TrainPage {
-  language: Locale
-  cancelled: string
-}
