--- packages/ui/.storybook/main.js
@@ -1,15 +1,0 @@
-module.exports = {
-  stories: [
-    '../components/**/*.stories.mdx',
-    '../components/**/*.stories.@(js|jsx|ts|tsx)'
-  ],
-  addons: [
-    '@storybook/addon-links',
-    '@storybook/addon-essentials',
-    '@storybook/addon-interactions'
-  ],
-  framework: '@storybook/react',
-  core: {
-    builder: 'webpack5'
-  }
-}
