--- packages/ui/.storybook/preview.js
@@ -1,3 +1,0 @@
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' }
-}
