+++ packages/ui/package.json
@@ -5,30 +5,13 @@
   "types": "dist/index.d.ts",
   "license": "MIT",
   "scripts": {
-    "build": "rimraf dist/* && tsc --build",
-    "test": "vitest run",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build": "tsc --build"
   },
   "dependencies": {
     "@junat/stitches": "*",
     "~digitraffic": "*"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.12",
-    "@storybook/addon-actions": "^6.5.0",
-    "@storybook/addon-essentials": "^6.5.0",
-    "@storybook/addon-interactions": "^6.5.0",
-    "@storybook/addon-links": "^6.5.0",
-    "@storybook/builder-webpack5": "^6.5.0",
-    "@storybook/manager-webpack5": "^6.5.0",
-    "@storybook/react": "^6.5.0",
-    "@storybook/testing-library": "^0.0.11",
-    "@types/react": "^18.0.9",
-    "babel-loader": "^8.2.5",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.6.4",
-    "vitest": "^0.12.6",
-    "webpack": "^5"
+    "typescript": "^4.6.4"
   }
 }
