+++ packages/ui/tsconfig.json
@@ -9,6 +9,5 @@
     "strict": true,
     "skipLibCheck": true,
     "outDir": "dist"
-  },
-  "exclude": ["**/**.stories.tsx", "tests", "dist"]
+  }
 }
