--- packages/ui/utils/get_hh_mm_time.ts
@@ -1,14 +1,0 @@
-/**
- * Formats a date into format `hh.mm` (hours.minutes).
- * @param {string} dateString
- * A valid date that can be parsed with `Date.parse()`
- * @see https://mdn.io/date_parse
- */
-export const getHhMmTime = (dateString: string) => {
-  const intl = Intl.DateTimeFormat('fi', {
-    hour: '2-digit',
-    minute: '2-digit'
-  })
-
-  return intl.format(Date.parse(dateString))
-}
