+++ site/src/pages/train/[date]/[trainNumber].tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/router'
 
 import { useMemo } from 'react'
 
-import { getTrainLongNames, getTrainPage } from '@junat/cms'
+import { getTrainLongNames } from '@junat/cms'
 
 import { SingleTimetable } from '@junat/ui'
 import useLiveTrain from '@hooks/use_live_train.hook'
@@ -24,14 +24,12 @@ interface TrainPageProps {
   longNames: TrainLongName[]
   trainNumber: number
   departureDate: string
-  cancelled: string
 }
 
 export default function TrainPage({
   longNames,
   trainNumber,
-  departureDate,
-  cancelled
+  departureDate
 }: TrainPageProps) {
   const [train, error] = useLiveTrain({
     trainNumber,
@@ -66,7 +64,6 @@ export default function TrainPage({
 
         {train && stations && (
           <SingleTimetable
-            cancelledText={cancelled}
             timetableRows={train.timeTableRows}
             locale={locale}
             stations={stations}
@@ -81,11 +78,7 @@ export default function TrainPage({
 TrainPage.layout = Page
 
 export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const locale = getLocaleOrThrow(context.locale)
-
-  const longNames = await getTrainLongNames(locale)
-  const { cancelled } = await getTrainPage(locale)
-
+  const longNames = await getTrainLongNames(getLocaleOrThrow(context.locale))
   const departureDate = context.query.date as unknown as string
   const trainNumber = context.query.trainNumber as unknown as string
 
@@ -98,8 +91,7 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     props: {
       longNames,
       trainNumber,
-      departureDate,
-      cancelled
+      departureDate
     }
   }
 }
